Add files via upload
The simplest way I have found to launch a Haiku desktop on a UEFI machine is to setup rEFInd on a USB drive as Partition 1, so it launches first, set the UEFI to launch UEFI USB devices first, and make sure I have a folder called Haiku under EFI with the file haiku_loader.efi I just updated. That will find any Haiku builds you have on your system, or at least it should. It doesn't always work, for one reason or another. But it's pretty darn good, and it's nice to have Haiku on the rEFInd menu at last!